# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/publishers": {
            "get": {
                "tags": [
                    "publishers"
                ],
                "summary": "Get list of publishers",
                "description": "Returns list of publishers",
                "operationId": "getpublishersList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PublisherResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "publishers"
                ],
                "summary": "Store new publisher",
                "description": "Returns publisher data",
                "operationId": "storePublisher",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePublisherRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PublisherResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/publishers/{id}": {
            "get": {
                "tags": [
                    "publishers"
                ],
                "summary": "Get publisher information",
                "description": "Returns publisher data",
                "operationId": "getPublisherById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Publisher id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PublisherResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "publishers"
                ],
                "summary": "Update existing publisher",
                "description": "Returns updated publisher data",
                "operationId": "updatePublisher",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Publisher id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePublisherRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PublisherResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "publishers"
                ],
                "summary": "Delete existing publisher",
                "description": "Deletes a record and returns no content",
                "operationId": "deletePublisher",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Publisher id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/admin/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get list of users",
                "description": "Returns list of users",
                "operationId": "getProfile",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserInfoResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "PublisherResource": {
                "title": "PublisherResource",
                "description": "Publisher resource",
                "properties": {
                    "publisher": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "ref": null
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PublisherResource"
                }
            },
...
            "StorePublisherRequest": {
                "title": "Store Publisher request",
                "description": "Store Publisher request body data",
                "required": [
                    "name, address, phone_number"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the new publisher",
                        "type": "string",
                        "example": "A nice publisher"
                    },
                    "address": {
                        "title": "address",
                        "description": "Address of the new publisher",
                        "type": "string",
                        "example": "A nice address"
                    },
                    "phone_number": {
                        "title": "phone_number",
                        "description": "Phone number of the new publisher",
                        "type": "string",
                        "example": "0123456789"
                    },
                    "": {
                        "title": "description",
                        "description": "Description of the new publisher",
                        "type": "string",
                        "example": "A nice description"
                    }
                },
                "type": "object"
            },
...
            "UpdatePublisherRequest": {
                "title": "Update Publisher request",
                "description": "Update Publisher request body data",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the new publisher",
                        "type": "string",
                        "example": "A nice publisher"
                    },
                    "address": {
                        "title": "address",
                        "description": "Address of the new publisher",
                        "type": "string",
                        "example": "A nice address"
                    },
                    "phone_number": {
                        "title": "phone_number",
                        "description": "Phone number of the new publisher",
                        "type": "string",
                        "example": "0123456789"
                    },
                    "description": {
                        "title": "description",
                        "description": "Description of the new publisher",
                        "type": "string",
                        "example": "A nice description"
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "bearer_token": []
        }
    ]
}